% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_shared_sp_matrix.R
\name{make_shared_sp_matrix}
\alias{make_shared_sp_matrix}
\title{Produce a "number of shared species" matrix from a presence/absence matrix}
\usage{
make_shared_sp_matrix(pres_abs_matrix)
}
\arguments{
\item{pres_abs_matrix}{a presence/absence matrix with sites as rows and species as columns}
}
\value{
matrix where each cell contains the number of species shared between the column and row.
}
\description{
Produce a "number of shared species" matrix from a presence/absence matrix
}
\details{
this might be useful for testing the reconstruction of a presence/absence matrix from a "shared species" matrix that includes alpha and pairwise beta values for sites.
}
