% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_across_sites_df.R
\name{opt_across_sites_df}
\alias{opt_across_sites_df}
\title{Select the sites that will maximize the number of present species}
\usage{
opt_across_sites_df(numbersites, data, destination = NULL)
}
\arguments{
\item{numbersites}{the number of sites to select}

\item{data}{a presence/absence matrix with sites as rows and species as columns}

\item{destination}{directory, including file name with .csv suffix, to which a .csv of solutions will be saved}
}
\value{
sorted list of selected sites with accompanying scores
}
\description{
Select the sites that will maximize the number of present species
}
\details{
This solves the maximum coverage problem. Given a matrix of presence/absence data for all sites, this picks the best combination of a user-defined-size subset of those sites to maximize the amount of "presence" across the selected sites.

WARNING: Because this considers every possible combination in a wildly inefficient way, it gets out-of-hand very quickly. For example, picking the best 6 sites out of 27 total requires ~296000 computations.
}
