% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_presabs_matrix.R
\name{produce_presabs_matrix}
\alias{produce_presabs_matrix}
\title{Produce a presence/absence matrix from a matrix of shared species between sites.}
\usage{
produce_presabs_matrix(sharedmatrix, nsamples = 1000)
}
\arguments{
\item{sharedmatrix}{a matrix where each cell contains the number of species shared between the column and row.}

\item{nsamples}{number of times to sample the possible new rows. Higher values take longer, but they are more likely to return the actual answers.}
}
\value{
Returns a presence/absence matrix with sites across the rows and species down the columns.
}
\description{
Produce a presence/absence matrix from a matrix of shared species between sites.
}
\details{
This function is designed to link alpha and beta diversity values to presence/absence for maximum coverage.

Each cell in the input matrix should be an integer number of species shared between sites. So the diagonal will be alpha diversities while other cells will represent a type of beta diversity.
}
