% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_newrow_presabs.R
\name{sample_newrow_presabs}
\alias{sample_newrow_presabs}
\title{Add a new row to a presence/absence matrix being reconstructed from a matrix of pairwise shared species}
\usage{
sample_newrow_presabs(sharedmatrix, nsamples = 10000, startingmatrix = NULL)
}
\arguments{
\item{sharedmatrix}{a matrix where each cell contains the number of species shared between the column and row.}

\item{nsamples}{number of times to sample the possible new rows. Higher values take longer, but they are more likely to return the actual answers.}

\item{startingmatrix}{the first few rows (two or more) of a presence/absence matrix being reconstructed. This might be, for example, the returned value from a previous run of this function.}
}
\value{
matrix (or matrices) of the input `startingmatrix` with an additional row of presence/absence data corresponding to the best scoring sample.
}
\description{
Add a new row to a presence/absence matrix being reconstructed from a matrix of pairwise shared species
}
\details{
This function is the opposite of `make_shared_sp_matrix()`, which converts a presence/absence table to a matrix of pairwise shared species.

`sample_newrow_presabs()` uses a matrix of pairwise shared values to add an additional row to a presence/absence matrix being constructed from those values. If no startingmatrix is provided, this function outputs the first three rows. Otherwise, this function adds a row to the input startingmatrix.

This function randomly samples possible new rows and scores each sample based on its distance from matching the matrix of pairwise shared species.
}
