% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_presabs_matrix.R
\name{simulate_presabs_matrix}
\alias{simulate_presabs_matrix}
\title{Produce a random presence/absence matrix of user-specified size}
\usage{
simulate_presabs_matrix(nrow, ncol)
}
\arguments{
\item{nrow}{the number of sites to select}

\item{ncol}{a presence/absence matrix with sites as rows and species as columns}
}
\value{
Returns a random presence/absence matrix of the specified size.
}
\description{
Produce a random presence/absence matrix of user-specified size
}
